# swagger.yaml
swagger: '2.0'
info:
  version: v4
  title: Projects API
host: 'localhost:3000'
basePath: /v4
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  '/projects/db':
    get:
      tags:
        - project
      operationId: findProjectsDB
      security:
        - Bearer: []
      description: Retrieve projects that match the filter directly from database
      responses:
        '200':
          description: A list of projects
          schema:
            $ref: '#/definitions/ProjectListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/limitParam'
        - name: filter
          required: true
          type: string
          in: query
          description: |
            Url encoded list of Supported filters
             - id
             - status
             - type
             - memberOnly
             - keyword
             - name
             - code
             - customer
             - manager
        - name: sort
          required: false
          description: >
            sort projects by status, name, type, createdAt, updatedAt. Default
            is createdAt asc
          in: query
          type: string
  '/projects':
    get:
      tags:
        - project
      operationId: findProjects
      security:
        - Bearer: []
      description: Retrieve projects that match the filter
      responses:
        '200':
          description: A list of projects
          schema:
            $ref: '#/definitions/ProjectListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/limitParam'
        - name: filter
          required: true
          type: string
          in: query
          description: |
            Url encoded list of Supported filters
             - id
             - status
             - type
             - memberOnly
             - keyword
             - name
             - code
             - customer
             - manager
        - name: sort
          required: false
          description: >
            sort projects by status, name, type, createdAt, updatedAt. Default
            is createdAt asc
          in: query
          type: string
    post:
      tags:
        - project
      operationId: addProject
      security:
        - Bearer: []
      description: Create a project
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/NewProjectBodyParam'
      responses:
        '201':
          description: Returns the newly created project
          schema:
            $ref: '#/definitions/ProjectResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /projects/metadata:
    get:
      tags:
        - metadata
      operationId: getAllMetadata
      security:
        - Bearer: []
      description: >-
        Retrieve all metadata including projectTemplates, productTemplates,
        milestoneTemplates, projectTypes, productCategories. All user roles can
        access this endpoint.
      responses:
        '200':
          description: The metadata
          schema:
            $ref: '#/definitions/AllMetadataResponse'
        '500':
          description: Invalid server state or unknown error
          schema:
            $ref: '#/definitions/ErrorModel'
  /projects/metadata/projectTemplates:
    get:
      tags:
        - projectTemplate
      operationId: findProjectTemplates
      security:
        - Bearer: []
      description: Retrieve all project templates. All user roles can access this endpoint.
      responses:
        '200':
          description: A list of project templates
          schema:
            $ref: '#/definitions/ProjectTemplateListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      tags:
        - projectTemplate
      operationId: addProjectTemplate
      security:
        - Bearer: []
      description: Create a project template
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ProjectTemplateBodyParam'
      responses:
        '201':
          description: Returns the newly created project template
          schema:
            $ref: '#/definitions/ProjectTemplateResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /projects/metadata/productTemplates:
    get:
      tags:
        - productTemplate
      operationId: findProductTemplates
      security:
        - Bearer: []
      description: Retrieve all product templates. All user roles can access this endpoint.
      responses:
        '200':
          description: A list of product templates
          schema:
            $ref: '#/definitions/ProductTemplateListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      tags:
        - productTemplate
      operationId: addProductTemplate
      security:
        - Bearer: []
      description: Create a product template
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ProductTemplateBodyParam'
      responses:
        '201':
          description: Returns the newly created product template
          schema:
            $ref: '#/definitions/ProductTemplateResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /projects/metadata/productCategories:
    get:
      tags:
        - productCategory
      operationId: findProductCategories
      security:
        - Bearer: []
      description: >-
        Retrieve all product categories. All user roles can access this
        endpoint.
      responses:
        '200':
          description: A list of product categories
          schema:
            $ref: '#/definitions/ProductCategoryListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      tags:
        - productCategory
      operationId: addProductCategory
      security:
        - Bearer: []
      description: >-
        Create a product category. Only admin or connect admin can access this
        endpoint.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ProductCategoryCreateBodyParam'
      responses:
        '201':
          description: Returns the newly created product category
          schema:
            $ref: '#/definitions/ProductCategoryResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
  '/projects/metadata/productCategories/{key}':
    get:
      tags:
        - productCategory
      description: >-
        Retrieve product category by id. All user roles can access this
        endpoint.
      security:
        - Bearer: []
      responses:
        '200':
          description: a product category
          schema:
            $ref: '#/definitions/ProductCategoryResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: '#/parameters/keyParam'
      operationId: getProductCategory
    patch:
      tags:
        - productCategory
      operationId: updateProductCategory
      security:
        - Bearer: []
      description: >-
        Update a product category. Only admin or connect admin can access this
        endpoint.
      responses:
        '200':
          description: Successfully updated product category.
          schema:
            $ref: '#/definitions/ProductCategoryResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: error payload
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: '#/parameters/keyParam'
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/ProductCategoryBodyParam'
    delete:
      tags:
        - productCategory
      description: >-
        Remove an existing product category. Only admin or connect admin can
        access this endpoint.
      security:
        - Bearer: []
      parameters:
        - $ref: '#/parameters/keyParam'
      responses:
        '204':
          description: Product category successfully removed
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: If product category is not found
          schema:
            $ref: '#/definitions/ErrorModel'
  /projects/metadata/projectTypes:
    get:
      tags:
        - projectType
      operationId: findProjectTypes
      security:
        - Bearer: []
      description: Retrieve all project types. All user roles can access this endpoint.
      responses:
        '200':
          description: A list of project types
          schema:
            $ref: '#/definitions/ProjectTypeListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      tags:
        - projectType
      operationId: addProjectType
      security:
        - Bearer: []
      description: >-
        Create a project type. Only admin or connect admin can access this
        endpoint.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/ProjectTypeCreateBodyParam'
      responses:
        '201':
          description: Returns the newly created project type
          schema:
            $ref: '#/definitions/ProjectTypeResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
  '/projects/metadata/projectTypes/{key}':
    get:
      tags:
        - projectType
      description: Retrieve project type by id. All user roles can access this endpoint.
      security:
        - Bearer: []
      responses:
        '200':
          description: a project type
          schema:
            $ref: '#/definitions/ProjectTypeResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: '#/parameters/keyParam'
      operationId: getProjectType
    patch:
      tags:
        - projectType
      operationId: updateProjectType
      security:
        - Bearer: []
      description: >-
        Update a project type. Only admin or connect admin can access this
        endpoint.
      responses:
        '200':
          description: Successfully updated project type.
          schema:
            $ref: '#/definitions/ProjectTypeResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: error payload
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: '#/parameters/keyParam'
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/ProjectTypeBodyParam'
    delete:
      tags:
        - projectType
      description: >-
        Remove an existing project type. Only admin or connect admin can access
        this endpoint.
      security:
        - Bearer: []
      parameters:
        - $ref: '#/parameters/keyParam'
      responses:
        '204':
          description: Project type successfully removed
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: If project is not found
          schema:
            $ref: '#/definitions/ErrorModel'
  /projects/metadata/orgConfig:
    get:
      tags:
        - orgConfig
      operationId: findOrgConfigs
      security:
        - Bearer: []
      description: >-
        Retrieve all organization configs. All user roles can access this
        endpoint.
      parameters:
        - name: filter
          required: true
          type: string
          in: query
          description: |
            Url encoded list of Supported filters
             - orgId (required)
             - configName
      responses:
        '200':
          description: A list of organization configs
          schema:
            $ref: '#/definitions/OrgConfigListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      tags:
        - orgConfig
      operationId: addOrgConfig
      security:
        - Bearer: []
      description: >-
        Create a organization config. Only admin or connect admin can access
        this endpoint.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/OrgConfigCreateBodyParam'
      responses:
        '201':
          description: Returns the newly created organization config
          schema:
            $ref: '#/definitions/OrgConfigResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /timelines:
    get:
      tags:
        - timeline
      operationId: findTimelines
      security:
        - Bearer: []
      description: Retrieve timelines which its projects are accessible by the user.
      parameters:
        - name: filter
          required: false
          type: string
          in: query
          description: |
            Url encoded list of supported filters
             - reference
             - referenceId
      responses:
        '200':
          description: A list of timelines
          schema:
            $ref: '#/definitions/TimelineListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      tags:
        - timeline
      operationId: addTimeline
      security:
        - Bearer: []
      description: >-
        Create a timeline. All users who can edit the project can access this
        endpoint.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/TimelineBodyParam'
      responses:
        '201':
          description: Returns the newly created timeline
          schema:
            $ref: '#/definitions/TimelineResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /timelines/metadata/milestoneTemplates:
    get:
      tags:
        - milestoneTemplates
      operationId: findMilestoneTemplates
      security:
        - Bearer: []
      description: >-
        Retrieve all milestone templates. All user roles can access this
        endpoint.
      parameters:
        - name: sort
          required: false
          description: sort by `order`. Default is `order asc`
          in: query
          type: string
        - name: filter
          required: false
          type: string
          in: query
          description: |
            Url encoded list of supported filters
             - reference
             - referenceId
      responses:
        '200':
          description: A list of milestone templates
          schema:
            $ref: '#/definitions/MilestoneTemplateListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      tags:
        - milestoneTemplates
      operationId: addMilestoneTemplate
      security:
        - Bearer: []
      description: >-
        Create a milestone template. Only connect manager, connect admin, and
        admin can access this endpoint. It also updates the `order` field of all
        other milestone templates in the same product template which have
        `order` greater than or equal to the `order` specified in the POST body.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/MilestoneTemplateBodyParam'
      responses:
        '201':
          description: Returns the newly created milestone template
          schema:
            $ref: '#/definitions/MilestoneTemplateResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
  /timelines/metadata/milestoneTemplates/clone:
    post:
      tags:
        - milestoneTemplates
      operationId: cloneMilestoneTemplate
      security:
        - Bearer: []
      description: >-
        Clone milestone templates from one product template to the other. Only
        connect manager, connect admin, and admin can access this endpoint.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/MilestoneCloneTemplateRequest'
      responses:
        '201':
          description: Returns the list of cloned milestone templates
          schema:
            $ref: '#/definitions/MilestoneTemplateListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
...
  '/projects/metadata/form/{key}':
    get:
      tags:
        - form version
      security:
        - Bearer: []
      description: get the latest revision of latest version for key.
      parameters:
        - $ref: '#/parameters/modelKeyParam'
      responses:
        '200':
          description: The model for the latest revision of latest version
          schema:
            $ref: '#/definitions/FormResponse'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: key not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Invalid server state or unknown error
          schema:
            $ref: '#/definitions/ErrorModel'
...
parameters:
...
  keyParam:
    name: key
    in: path
    description: project type key
    required: true
    type: string
...
  offsetParam:
    name: offset
    description: number of items to skip. Defaults to 0
    in: query
    required: false
    type: integer
    format: int32
  limitParam:
    name: limit
    description: max records to return. Defaults to 20
    in: query
    required: false
    type: integer
    format: int32
...
definitions:
...
  ErrorModel:
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            description: http status code
            type: integer
            format: int32
          debug:
            type: object
          content:
            type: object
...
  NewProjectBodyParam:
    type: object
    properties:
      param:
        $ref: '#/definitions/NewProject'
...
  ProjectListResponse:
    title: List response
    type: object
    properties:
      id:
        type: string
        readOnly: true
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: integer
            format: int32
            description: http status code
          metadata:
            $ref: '#/definitions/ResponseMetadata'
          content:
            type: array
            items:
              $ref: '#/definitions/Project'
...